@@ -40,9 +40,9 @@
     "@babel/preset-env": "^7.8.3",
     "@babel/preset-react": "^7.8.3",
     "@babel/types": "^7.8.3",
-    "@material-ui/core": "^3.8.1",
-    "@storybook/addon-actions": "^5.3.3",
-    "@storybook/react": "^5.3.3",
+    "@mui/material": "^5.5.0",
+    "@storybook/addon-actions": "^6.4.20",
+    "@storybook/react": "^6.4.20",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "eslint": "^6.8.0",
@@ -53,18 +53,18 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-standard": "^4.0.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0"
   },
   "dependencies": {
-    "@material-ui/styles": "^4.8.2",
+    "@mui/styles": "^5.5.0",
     "prop-types": "^15.7.2",
     "react-color": "^2.18.0",
     "recompose": "^0.30.0"
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.8.3",
-    "react": "^16.12.0"
+    "@mui/material": "^5.5.0",
+    "react": "^17.0.0"
   },
   "types": "src/index.d.ts"
 }
